more fixes on RESTEASY-794
